requirements.txt
accelerate
affine
albumentations
annotated-types
appdirs
appnope
asttokens
async-timeout
attrs
backcall
beautifulsoup4
bleach
certifi
charset-normalizer
click
click-plugins
cligj
cloudpickle
configilm
decorator
defusedxml
docker-pycreds
docopt
executing
fastjsonschema
filelock
frozenlist
fsspec
future
gitdb
GitPython
gym
gym-notices
huggingface-hub
hyperopt
idna
ipython
jedi
Jinja2
joblib
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
jupyterlab_pygments
lightning
lightning-utilities
lmdb
MarkupSafe
matplotlib-inline
mistune
more-itertools
mpmath
multidict
nbclient
nbconvert
nbformat
networkx
numpy
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
peft==0.15.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.9
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
rasterio==1.4.3
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.20.0
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentry-sdk==2.18.0
setproctitle==1.3.3
simsimd==6.2.1
six==1.16.0
smmap==5.0.1
soupsieve==2.6
stack-data==0.6.3
stringzilla==3.12.3
sympy==1.13.1
threadpoolctl==3.5.0
timm==0.9.16
tinycss2==1.3.0
tokenizers==0.21.0
torch==2.5.1
torchmetrics==1.6.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.0
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.18.6
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
yarl==1.18.3
